Spring integration started.
