--- src/main/java/com/burachenko/munichhotel/configuration/VaadinConfiguration.java
@@ -1,8 +1,0 @@
-package com.burachenko.munichhotel.configuration;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class VaadinConfiguration {
-
-}
