+++ src/main/java/com/burachenko/munichhotel/ui/view/AbstractEntityView.java
@@ -17,8 +17,7 @@ import com.vaadin.ui.HorizontalLayout;
 import com.vaadin.ui.TextField;
 import com.vaadin.ui.VerticalLayout;
 import com.vaadin.ui.Window;
-import com.vaadin.ui.components.grid.MultiSelectionModel;
-import com.vaadin.ui.components.grid.MultiSelectionModel.SelectAllCheckBoxVisibility;
+import org.springframework.beans.factory.annotation.Autowired;
 
 abstract class AbstractEntityView<DTO extends AbstractDto, Service extends AbstractService> extends VerticalLayout implements View {
 
@@ -35,7 +34,9 @@ abstract class AbstractEntityView<DTO extends AbstractDto, Service extends Abstr
 
     private DataProvider<DTO, String> dataProvider;
     private ConfigurableFilterDataProvider<DTO, Void, String> filteredDataProvider;
-    private Grid<DTO> grid = new Grid<>(getDtoClass());
+
+    @Autowired
+    private Grid<DTO> grid;
 
     public AbstractEntityView(final Service service) {
         this.service = service;
@@ -100,10 +101,6 @@ abstract class AbstractEntityView<DTO extends AbstractDto, Service extends Abstr
 
     private void setupGrid(){
         setupGridDataProvider();
-        grid.setSelectionMode(Grid.SelectionMode.MULTI);
-        ((MultiSelectionModel) grid.getSelectionModel())
-                .setSelectAllCheckBoxVisibility(SelectAllCheckBoxVisibility.VISIBLE);
-        grid.setSizeFull();
         grid.addSelectionListener(selection -> {
             final int selectedLinesCount = selection.getAllSelectedItems().size();
             if (selectedLinesCount == 1){
@@ -124,11 +121,6 @@ abstract class AbstractEntityView<DTO extends AbstractDto, Service extends Abstr
                 query -> service.findByFilterQueryWithPagination(query),
                 query -> (int) service.findByFilterQueryWithPagination(query).count()
         );
-
-//        filteredDataProvider = DataProvider.fromFilteringCallbacks(
-//                query -> service.findByFilterQueryWithPagination(query),
-//                query -> (int) service.findByFilterQueryWithPagination(query).count()
-//        ).withConfigurableFilter();
         filteredDataProvider = dataProvider.withConfigurableFilter();
         grid.setDataProvider(filteredDataProvider);
         grid.getDataProvider().refreshAll();
